+++ tests/testthat.R
@@ -1,4 +1,10 @@
 library(testthat)
+
+if (Sys.getenv("TENSORFLOW_EAGER") == "TRUE")
+  tensorflow::tfe_enable_eager_execution()
+
 library(keras)
 
-test_check("keras")
+if (identical(Sys.getenv("NOT_CRAN"), "true")) {
+  test_check("keras") 
+}