+++ tests/testthat/test-model-persistence.R
@@ -1,5 +1,5 @@
 
-context("model persistence")
+context("model-persistence")
 
 source("utils.R")
 
@@ -45,10 +45,9 @@ test_succeeds("model with custom loss and metrics can be saved and loaded", {
   labels <- matrix(round(runif(1000*10, min = 0, max = 9)), nrow = 1000, ncol = 10)
   
   
-  model %>% fit(data, labels, epochs = 2)
+  model %>% fit(data, labels, epochs = 2, verbose = 0)
   
 })
-
 
 test_succeeds("model weights can be saved and loaded", {
 
@@ -92,14 +91,14 @@ test_succeeds("model can be saved and loaded from R 'raw' object", {
 })
 
 test_succeeds("saved models/weights are mirrored in the run_dir", {
-  run <- tfruns::training_run("train.R")
+  run <- tfruns::training_run("train.R", echo = FALSE)
   run_dir <- run$run_dir
   expect_true(file.exists(file.path(run_dir, "model.h5")))
   expect_true(file.exists(file.path(run_dir, "weights", "weights.h5")))
 })
 
 test_succeeds("callback output is redirected to run_dir", {
-  run <- tfruns::training_run("train.R")
+  run <- tfruns::training_run("train.R", echo = FALSE)
   run_dir <- run$run_dir
   if (is_backend("tensorflow"))
     expect_true(file_test("-d", file.path(run_dir, "tflogs")))
@@ -109,18 +108,56 @@ test_succeeds("callback output is redirected to run_dir", {
 
 test_succeeds("model can be exported to TensorFlow", {
   if (!is_backend("tensorflow")) skip("not a tensorflow backend")
-
+  
   model <- define_and_compile_model()
   model_dir <- tempfile()
   
   export <- function() tensorflow::export_savedmodel(model, model_dir)
   
-  if (!grepl("^keras", Sys.getenv("KERAS_IMPLEMENTATION"))) {
-    expect_error(export())
-  }
-  else {
-    export()
-    model_files <- dir(model_dir, recursive = TRUE)
-    expect_true(any(grepl("saved_model\\.pb", model_files)))
-  }
+  export()
+  model_files <- dir(model_dir, recursive = TRUE)
+  expect_true(any(grepl("saved_model\\.pb", model_files)))
+
 })
+
+test_succeeds("model can be exported to saved model format", {
+  if (!is_backend("tensorflow")) skip("not a tensorflow backend")
+  if (!tensorflow::tf_version() >= "1.14") skip("Needs TF >= 1.14")
+  
+  model <- define_and_compile_model()
+  data <- matrix(rexp(1000*784), nrow = 1000, ncol = 784)
+  labels <- matrix(round(runif(1000*10, min = 0, max = 9)), nrow = 1000, ncol = 10)
+  
+  model %>% fit(data, labels, epochs = 2, verbose = 0)
+  
+  model_dir <- tempfile()
+  dir.create(model_dir)
+  
+  model_to_saved_model(model, model_dir)
+  loaded <- model_from_saved_model(model_dir)
+  
+  expect_equal(
+    predict(model, matrix(rep(1, 784), nrow = 1)),
+    predict(loaded, matrix(rep(1, 784), nrow = 1))
+  )
+})
+
+test_succeeds("model can be exported to saved model format using save_model_tf", {
+  
+  if (!is_backend("tensorflow")) skip("not a tensorflow backend")
+  if (!tensorflow::tf_version() >= "2.0.0") skip("Needs TF >= 2.0")
+  
+  model <- define_and_compile_model()
+  model_dir <- tempfile()
+  
+  s <- save_model_tf(model, model_dir)
+  loaded <- load_model_tf(model_dir)
+  
+  expect_equal(
+    predict(model, matrix(rep(1, 784), nrow = 1)),
+    predict(loaded, matrix(rep(1, 784), nrow = 1))
+  )
+})
+
+
+